pkg/homedir: unify implementations
Unify some of the exported functions instead of maintaining separate functions
per platform.
